refactor(pages/auth): group auth flows into FSD-style slices
