+++ src/pages/auth/signup/index.ts
