+++ src/pages/auth/signup/model/schemas.ts
@@ -1,5 +1,5 @@
 import { z } from 'zod/v4';
-import { SAuth } from 'entities/auth';
+import { CAuth, SAuth } from 'entities/auth';
 
 const MAX_FULL_NAME_WORDS = 2;
 
@@ -9,8 +9,8 @@ export const SignupForm = z
       .string()
       .trim()
       .min(1, 'Обязательное поле')
-      .min(2, 'Слишком короткое имя')
-      .max(100, 'Слишком длинное имя')
+      .min(CAuth.MIN_NAME_LENGTH, 'Слишком короткое имя')
+      .max(CAuth.MAX_NAME_LENGTH, 'Слишком длинное имя')
       .refine((name) => name.split(/\s+/).length <= MAX_FULL_NAME_WORDS, {
         message: 'Введите только имя и фамилию',
       }),
