+++ src/pages/auth/signup/model/types.ts
