+++ src/pages/auth/signup/model/utils/prepare-fullname.ts
